requirements.txt
flask
flask-cors
langchain
langchain-community
chromadb
openai
python-dotenv

-e .